# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Fellow traveler Application",
        "title": "Fellow traveler API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8484",
    "basePath": "/",
    "paths": {
        "/api/orders": {
            "get": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "get list of all orders",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get All Orders",
                "operationId": "get-all-orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "fill if you need search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "fill if you need search",
                        "name": "is_response",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "fill if you need search",
                        "name": "all_price",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GetOrder"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "create new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create Order",
                "operationId": "create-order",
                "parameters": [
                    {
                        "description": "new order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.defaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "get order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get Order By ID",
                "operationId": "get-order-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the route",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetOrder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/report": {
            "get": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "get list of report",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Report",
                "operationId": "get-report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "fill if you need search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "fill if you need search",
                        "name": "is_response",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "fill if you need search",
                        "name": "all_price",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Reports"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/routes/{id}": {
            "put": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "update existed order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update Order",
                "operationId": "update-order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the order",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order update info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.defaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "delete route by ID",
                "tags": [
                    "orders"
                ],
                "summary": "Delete Order By ID",
                "operationId": "delete-route-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the route",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.defaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AKA": []
                    }
                ],
                "description": "check as response existed order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Check order as response",
                "operationId": "check-order-as-response",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the route",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " check order as response info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Checkresponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.defaultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "controllers.defaultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Checkresponse": {
            "type": "object",
            "properties": {
                "is_response": {
                    "type": "boolean"
                }
            }
        },
        "models.GetOrder": {
            "type": "object",
            "properties": {
                "all_price": {
                    "type": "integer"
                },
                "client_phone": {
                    "type": "string"
                },
                "driver_phone": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "into": {
                    "type": "string"
                },
                "is_response": {
                    "type": "boolean"
                },
                "start_price": {
                    "type": "integer"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "into": {
                    "type": "string"
                }
            }
        },
...
        "models.Reports": {
            "type": "object",
            "properties": {
                "all_price": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "integer"
                },
                "distance": {
                    "type": "integer"
                },
                "driver_id": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "into": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_response": {
                    "type": "boolean"
                },
                "start_price": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AKA": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}